+++ src/main/java/frc/robot/subsystems/swervedrive/SwerveSubsystem.java
@@ -658,6 +658,16 @@ public class SwerveSubsystem extends SubsystemBase {
     }
   }
 
+  public void zeroGyroAgainstAlliance() {
+    if (!isRedAlliance()) {
+      zeroGyro();
+      // Set the pose 180 degrees
+      //resetOdometry(new Pose2d(getPose().getTranslation(), Rotation2d.fromDegrees(180)));
+    } else {
+      zeroGyro();
+      resetOdometry(new Pose2d(getPose().getTranslation(), Rotation2d.fromDegrees(180)));
+    }
+  }
   /**
    * Sets the drive motors to brake/coast mode.
    *
